+++ src/Http/Middleware/EnsureHasRole.php
@@ -5,9 +5,27 @@ namespace BSPDX\Keystone\Http\Middleware;
 use Closure;
 use Illuminate\Http\Request;
 use Symfony\Component\HttpFoundation\Response;
+use BSPDX\Keystone\Services\Contracts\AuthorizationServiceInterface;
 
 class EnsureHasRole
 {
+    /**
+     * The authorization service instance.
+     *
+     * @var AuthorizationServiceInterface
+     */
+    protected AuthorizationServiceInterface $authorizationService;
+
+    /**
+     * Create a new middleware instance.
+     *
+     * @param AuthorizationServiceInterface $authorizationService
+     */
+    public function __construct(AuthorizationServiceInterface $authorizationService)
+    {
+        $this->authorizationService = $authorizationService;
+    }
+
     /**
      * Handle an incoming request.
      *
@@ -21,11 +39,11 @@ class EnsureHasRole
         }
 
         // Super admins bypass all role checks
-        if (method_exists($request->user(), 'isSuperAdmin') && $request->user()->isSuperAdmin()) {
+        if ($this->authorizationService->userCanBypassPermissions($request->user())) {
             return $next($request);
         }
 
-        if (!$request->user()->hasAnyRole($roles)) {
+        if (!$this->authorizationService->userHasAnyRole($request->user(), $roles)) {
             abort(403, 'You do not have the required role to access this resource.');
         }